pyproject.toml
[project]
name = "travel-genie"
version = "0.1.0"
description = "AI Travel Recommendation System"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "fastmcp>=0.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=core",
    "--cov=tools",
    "--cov=agent",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["core", "tools", "agent"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
